sql/row: move the table lookup logic to a separate file; add comments
Release note: None
