+++ pkg/sql/row/fk_existence_prepare.go
@@ -17,152 +17,12 @@ package row
 import (
 	"context"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 )
 
-// ID is an alias for sqlbase.ID.
-type ID = sqlbase.ID
-
-// TableLookupsByID maps table IDs to looked up descriptors or, for tables that
-// exist but are not yet public/leasable, entries with just the IsAdding flag.
-type TableLookupsByID map[ID]TableLookup
-
-// TableLookup is the value type of TableLookupsByID: An optional table
-// descriptor, populated when the table is public/leasable, and an IsAdding
-// flag.
-// This also includes an optional CheckHelper for the table.
-type TableLookup struct {
-	Table       *sqlbase.ImmutableTableDescriptor
-	IsAdding    bool
-	CheckHelper *sqlbase.CheckHelper
-}
-
-// TableLookupFunction is the function type used by TablesNeededForFKs that will
-// perform the actual lookup.
-type TableLookupFunction func(context.Context, ID) (TableLookup, error)
-
-// NoLookup can be used to not perform any lookups during a TablesNeededForFKs
-// function call.
-func NoLookup(_ context.Context, _ ID) (TableLookup, error) {
-	return TableLookup{}, nil
-}
-
-// CheckPrivilegeFunction is the function type used by TablesNeededForFKs that will
-// check the privileges of the current user to access specific tables.
-type CheckPrivilegeFunction func(context.Context, sqlbase.DescriptorProto, privilege.Kind) error
-
-// NoCheckPrivilege can be used to not perform any privilege checks during a
-// TablesNeededForFKs function call.
-func NoCheckPrivilege(_ context.Context, _ sqlbase.DescriptorProto, _ privilege.Kind) error {
-	return nil
-}
-
-// FKCheck indicates a kind of FK check (delete, insert, or both).
-type FKCheck int
-
-const (
-	// CheckDeletes checks if rows reference a changed value.
-	CheckDeletes FKCheck = iota
-	// CheckInserts checks if a new value references an existing row.
-	CheckInserts
-	// CheckUpdates checks all references (CheckDeletes+CheckInserts).
-	CheckUpdates
-)
-
-type tableLookupQueueElement struct {
-	tableLookup TableLookup
-	usage       FKCheck
-}
-
-type tableLookupQueue struct {
-	queue          []tableLookupQueueElement
-	alreadyChecked map[ID]map[FKCheck]struct{}
-	tableLookups   TableLookupsByID
-	lookup         TableLookupFunction
-	checkPrivilege CheckPrivilegeFunction
-	analyzeExpr    sqlbase.AnalyzeExprFunction
-}
-
-func (tl *TableLookup) addCheckHelper(
-	ctx context.Context, analyzeExpr sqlbase.AnalyzeExprFunction,
-) error {
-	if analyzeExpr == nil {
-		return nil
-	}
-	tableName := tree.MakeUnqualifiedTableName(tree.Name(tl.Table.Name))
-	tl.CheckHelper = &sqlbase.CheckHelper{}
-	return tl.CheckHelper.Init(ctx, analyzeExpr, &tableName, tl.Table)
-}
-
-func (q *tableLookupQueue) getTable(ctx context.Context, tableID ID) (TableLookup, error) {
-	if tableLookup, exists := q.tableLookups[tableID]; exists {
-		return tableLookup, nil
-	}
-	tableLookup, err := q.lookup(ctx, tableID)
-	if err != nil {
-		return TableLookup{}, err
-	}
-	if !tableLookup.IsAdding && tableLookup.Table != nil {
-		if err := q.checkPrivilege(ctx, tableLookup.Table, privilege.SELECT); err != nil {
-			return TableLookup{}, err
-		}
-		if err := tableLookup.addCheckHelper(ctx, q.analyzeExpr); err != nil {
-			return TableLookup{}, err
-		}
-	}
-	q.tableLookups[tableID] = tableLookup
-	return tableLookup, nil
-}
-
-func (q *tableLookupQueue) enqueue(ctx context.Context, tableID ID, usage FKCheck) error {
-	// Lookup the table.
-	tableLookup, err := q.getTable(ctx, tableID)
-	if err != nil {
-		return err
-	}
-	// Don't enqueue if lookup returns an empty tableLookup. This just means that
-	// there is no need to walk any further.
-	if tableLookup.Table == nil {
-		return nil
-	}
-	// Only enqueue checks that haven't been performed yet.
-	if alreadyCheckByTableID, exists := q.alreadyChecked[tableID]; exists {
-		if _, existsInner := alreadyCheckByTableID[usage]; existsInner {
-			return nil
-		}
-	} else {
-		q.alreadyChecked[tableID] = make(map[FKCheck]struct{})
-	}
-	q.alreadyChecked[tableID][usage] = struct{}{}
-	// If the table is being added, there's no need to check it.
-	if tableLookup.IsAdding {
-		return nil
-	}
-	switch usage {
-	// Insert has already been checked when the table is fetched.
-	case CheckDeletes:
-		if err := q.checkPrivilege(ctx, tableLookup.Table, privilege.DELETE); err != nil {
-			return err
-		}
-	case CheckUpdates:
-		if err := q.checkPrivilege(ctx, tableLookup.Table, privilege.UPDATE); err != nil {
-			return err
-		}
-	}
-	(*q).queue = append((*q).queue, tableLookupQueueElement{tableLookup: tableLookup, usage: usage})
-	return nil
-}
-
-func (q *tableLookupQueue) dequeue() (TableLookup, FKCheck, bool) {
-	if len((*q).queue) == 0 {
-		return TableLookup{}, 0, false
-	}
-	elem := (*q).queue[0]
-	(*q).queue = (*q).queue[1:]
-	return elem.tableLookup, elem.usage, true
-}
+// The utilities in this file facilitate the prepare phase of FK
+// existence checks, during or at the end of logical planning, and
+// before execution starts.
 
 // TablesNeededForFKs populates a map of TableLookupsByID for all the
 // TableDescriptors that might be needed when performing FK checking for delete
